# trace: optimize span id parsing