Decreases runtime by about 2.6x (9.5 ns to 3.7 ns) for parsing the 16-char hex string.

```sh
BenchmarkParseUint64
BenchmarkParseUint64/ParseUint64Table
BenchmarkParseUint64/ParseUint64Table-12         	318850330	         3.686 ns/op	       0 B/op	       0 allocs/op
BenchmarkParseUint64/ParseUint64
BenchmarkParseUint64/ParseUint64-12              	126126168	         9.553 ns/op	       0 B/op	       0 allocs/op
BenchmarkParseUint64/hex.DecodeString
BenchmarkParseUint64/hex.DecodeString-12         	69341446	        16.55 ns/op	       8 B/op	       1 allocs/op
```